Github action to build with dune package management
This adds a dune lockdir to the project and a github action where
pushing a tag causes a vpn binary distribution to be built for linux and
macos, and released to the github "releases" page.

Also fixes a few minor issues which prevented the project from building
with the locked versions of dependencies.

Signed-off-by: Stephen Sherratt <stephen@sherra.tt>